Rename date.rb to prevent bundler issues
