+++ bank_holiday.gemspec
@@ -5,7 +5,7 @@ require 'bank_holiday/version'
 Gem::Specification.new do |s|
   s.name        = 'bank_holiday'
   s.version     = BankHoliday::VERSION
-  s.date        = '2014-07-25'
+  s.date        = '2017-11-06'
   s.summary     = "Finding bank holidays including helpers for dates."
   s.description = "Finding bank holidays including helpers for dates."
   s.authors     = ["Craig Sheen"]
@@ -13,7 +13,7 @@ Gem::Specification.new do |s|
   s.files       = ["lib/bank_holiday.rb"]
   s.homepage    = 'https://github.com/craigsheen/bank_holiday'
   s.license     = 'MIT'
-  
+
   s.add_dependency 'dish'
   s.require_paths = ["lib", "lib/bank_holiday"]
   s.files         = `git ls-files`.split($/)
