+++ lib/bank_holiday.rb
@@ -1,8 +1,9 @@
 require 'net/http'
 require 'dish/ext'
-require 'bank_holiday/date'
+require 'date'
+require 'bank_holiday/date_extension'
 require 'bank_holiday/bank_holiday'
 
 module BankHoliday
-  
+
 end
