+++ lib/bank_holiday/date_extension.rb
