+++ lib/bank_holiday/version.rb
@@ -1,3 +1,3 @@
 module BankHoliday
-  VERSION = "0.0.3"
+  VERSION = "0.0.4"
 end
